+++ bin/test.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+sudo python3 setup.py install
+sudo onlysnarf -debug -show